+++ src/Telegram.Bot/Requests/Available Methods/Admin Bot/UnpinChatMessageRequest.cs
@@ -1,0 +1,30 @@
+﻿using Newtonsoft.Json;
+using Newtonsoft.Json.Serialization;
+using Telegram.Bot.Types;
+
+// ReSharper disable once CheckNamespace
+namespace Telegram.Bot.Requests
+{
+    /// <summary>
+    /// Unpin a message in a supergroup or a channel. The bot must be an administrator in the chat for this to work and must have the ‘can_pin_messages’ admin right in the supergroup or ‘can_edit_messages’ admin right in the channel.
+    /// </summary>
+    [JsonObject(MemberSerialization.OptIn, NamingStrategyType = typeof(SnakeCaseNamingStrategy))]
+    public class UnpinChatMessageRequest : RequestBase<bool>
+    {
+        /// <summary>
+        /// Unique identifier for the target chat or username of the target channel (in the format @channelusername)
+        /// </summary>
+        [JsonProperty(Required = Required.Always)]
+        public ChatId ChatId { get; set; }
+
+        /// <summary>
+        /// Initializes a new request with chatId
+        /// </summary>
+        /// <param name="chatId">Unique identifier for the target chat or username of the target channel</param>
+        public UnpinChatMessageRequest(ChatId chatId)
+            : base("unpinChatMessage")
+        {
+            ChatId = chatId;
+        }
+    }
+}