+++ test/Telegram.Bot.Tests.Integ/Common/Fixtures/EntityFixture.cs
@@ -1,7 +1,0 @@
-﻿namespace Telegram.Bot.Tests.Integ.Common.Fixtures
-{
-    public class EntityFixture<TEntity>
-    {
-        public TEntity Entity { get; set; }
-    }
-}