+++ test/Telegram.Bot.Tests.Integ/Location/InlineMessageLiveLocationTests .cs
@@ -1,6 +1,6 @@
 using System;
 using System.Threading.Tasks;
-using Telegram.Bot.Tests.Integ.Common;
+using Telegram.Bot.Tests.Integ.Framework;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.InlineKeyboardButtons;
 using Telegram.Bot.Types.InlineQueryResults;