+++ test/Telegram.Bot.Tests.Integ/Location/LiveLocationTests.cs
@@ -1,7 +1,7 @@
 using System.Threading.Tasks;
 using Newtonsoft.Json.Linq;
-using Telegram.Bot.Tests.Integ.Common;
-using Telegram.Bot.Tests.Integ.Common.Fixtures;
+using Telegram.Bot.Tests.Integ.Framework;
+using Telegram.Bot.Tests.Integ.Framework.Fixtures;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 using Xunit;