+++ test/Telegram.Bot.Tests.Integ/Other/ChatInfoTests.cs
@@ -1,6 +1,6 @@
 ﻿using System.Threading.Tasks;
-using Telegram.Bot.Tests.Integ.Common;
-using Telegram.Bot.Tests.Integ.Common.Fixtures;
+using Telegram.Bot.Tests.Integ.Framework;
+using Telegram.Bot.Tests.Integ.Framework.Fixtures;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 using Xunit;