+++ test/Telegram.Bot.Tests.Integ/Sending Messages/AlbumMessageTests.cs
@@ -2,8 +2,8 @@
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using Telegram.Bot.Tests.Integ.Common;
-using Telegram.Bot.Tests.Integ.Common.Fixtures;
+using Telegram.Bot.Tests.Integ.Framework;
+using Telegram.Bot.Tests.Integ.Framework.Fixtures;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 using Xunit;