+++ test/Telegram.Bot.Tests.Integ/Sending Messages/MessageReplyMarkupTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.Linq;
 using System.Threading.Tasks;
-using Telegram.Bot.Tests.Integ.Common;
-using Telegram.Bot.Tests.Integ.Common.Fixtures;
+using Telegram.Bot.Tests.Integ.Framework;
+using Telegram.Bot.Tests.Integ.Framework.Fixtures;
 using Telegram.Bot.Types;
 using Telegram.Bot.Types.Enums;
 using Telegram.Bot.Types.ReplyMarkups;