+++ test/Telegram.Bot.Tests.Integ/Stickers/StickersTestsFixture.cs
@@ -1,5 +1,5 @@
 ﻿using System.Collections.Generic;
-using Telegram.Bot.Tests.Integ.Common;
+using Telegram.Bot.Tests.Integ.Framework;
 using Telegram.Bot.Types;
 
 namespace Telegram.Bot.Tests.Integ.Stickers